Added .gitignore & fixed compiler warning about unsigned compares

+++ main.cpp
@@ -9,7 +9,7 @@ using namespace std;
 int main(){
     Game g = Game();
 
-    for(int i = 0; i < g.p1.hand.size(); i++){
+    for(size_t i = 0; i < g.p1.hand.size(); i++){
         cout << g.p1.hand.at(i).getColor() << " " << g.p1.hand.at(i).getValue();
         if(g.p1.hand.at(i).getIsAction() == true){
             cout << " " << g.p1.hand.at(i).getAction() << endl;
